+++ src/Akka.Core/Util/Internal/Extensions.cs
@@ -65,7 +65,7 @@ namespace Akka.Util.Internal
             var j = 0;
             while (true)
             {
-                if (j > path.Length) yield break;
+                if ((uint)j > (uint)path.Length) yield break;
                 else if (path[j] == '\"')
                 {
                     i = path.IndexOf('\"', j + 1);
